// libs/features/benefits/src/lib/shared/models/enable-repayment-response.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
import { BilingualText } from '@gosi-ui/core';

export class EnableRepaymentResponse {
  benefitRequestId: number;
  enableLumpsumRepaymentId: number;
  message: BilingualText;
}